+++ vast/integration/integration.py
@@ -316,7 +316,7 @@ class Server:
         self.name = name
         self.cwd = work_dir / self.name
         self.port = port
-        command = [self.app]
+        command = [self.app, "--disable-default-config-dirs"]
         if self.config_arg:
             command.append(self.config_arg)
         command = command + args